Separate plugin loading and initialization
Separate plugin loading and initialization so that the list of available plugins
is fixed at start-up and easily accessible without necessarily causing them to
be initialized. This also allows the list of plugin classes to be accessible
before the Moksha Hub configuration is even available. In addition, plugin
entry-points are now required to either be a specific plugin class or collection
of them. There is an associated commit for statscache_plugins.