Dynamically generate producers based on plugins
Replaces the static list of available producers with a class factory generator
that determines the frequencies desired by the available plugin classes and then
creates a producer class for each frequency. As part of this, producers are now
responsible for initializing their respective plugins and registering them in the
plugin table for access by the web app. This will facilitate merging the
initialize() method of plugins with the standard __init__(), since a database
session will now be available at the time of plugin instantiation.