Synchronize producer/plugin frequencies
The statscache.plugins.Frequency class is repurposed as
statscache.frequency.Frequency to serve as a sort of synchronized timedelta.
The producer factory now initializes all instances of the class so that they
are all synchronized on the same moment, UTC midnight of the current day.
To faciltate this, plugin classes now include a timedelta instance as an
'interval' class attribute and receive the associated Frequency instance in the
__init__() method, which attaches it as a 'frequency' instance attribute.