Include a default value for frequency epoch
Default the synchronization datetime, or epoch, of statscache.frequency.Frequency
instances to UTC midnight of the current day that the module was instantiated,
making it easier for future uses of the class to get consistent behavior. The
only place where objects of the class are currently created, in
statscache.producer.factory(), is switched to the default value, which results
in no behavioral change.