Separate frontend plugin initialization (bugfix)
The frontend and backend of statscache share very little in the nature of how
they initialize and access plugin instances, especially with the recent addition
of the producer factory. In fact, their requirements for initialization are
largely incompatible, and running the frontend in its own process did not work,
as plugin instantiation was tied to producer initialization. This is a bug. To
remedy this, the common plugin table is done away with in favor of totally
separate plugin initialization and management paths for the backend and frontend.
For the most part, all this entailed was implementing plugin initialization and
management directly in the web interface code paths.