Restructure plugin system
This commit implements a new internal architecture for statscache. Incoming
fedmsg traffic is no longer buffered for any period; instead, plugins process
each message immediately when received by the stats consumer. To avoid the
potential bottleneck of database I/O, message processing and model updates are
disasscoiated, so that each plugin internally accumulates the data needed to
update its model as a result of the message processing. Specifically, each
plugin has two new methods: process(), to be called immediately on the receipt
of a message, and update(), to be called at a regular interval (set by the
active statscache configuration) to update the plugin's model.

This new internal architecture simplifies many aspects of statscache related to
plugins and has allowed for an assortment of unrelated code refactoring and
reductions.