Conceal Twisted's callback composition
In Twisted, the return value of one callback is given to the next as its input,
instead of the value which resolved the Deferred instance. As this usage pattern
does not fit with the simplistic model targeted by the statscache plugin API, it
is abstracted away by the interface exposed in statscache.plugins.threads.