Add start/stop datetime-pickers to model feed view
The start datetime defaults to the epoch, and the stop datetime defaults to
moment at which the page was generated. Both of these times are normalized to
the next valid minute, in order to suppress the datetime picker widgets from
doing so, which triggers and unnecessary reload of the model content.

Additionally, the start datetime-picker disallows selection of datetimes before
the epoch and after the currently chosen stop datetime, with the stop datetime-
picker supporting similar behavior with respect to the current moment and the
currently chosen start datetime.

Some changes to the identification of the model table were also necessary, in
order to prevent the web page's JavaScript from interfering with the datetime-
pickers' elements.