Fix plugin_model() handling of query argument
In statscache.app.plugin_model(), the limit() method was being called on the
relevant SQLAlchemy queryset with the intention of limiting the *entire* query,
such that future manipulation of the queryset would behave as if the selected
model only had that many rows. In reality, the necessary manipulation of the
queryset in the statscache.app.paginate() function would, in many cases,
circumvent this value unintentionally.